requirements.txt
applicationinsights==0.11.9
atomicwrites==1.3.0
attrs==19.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-auth==1.6.3
idna==2.8
importlib-metadata==0.18
kubernetes==9.0.0
more-itertools==7.0.0
oauthlib==3.0.1
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyparsing==2.4.0
pytest==4.6.3
pytest-mock==1.10.4
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.yaml==0.15.97
six==1.12.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
zipp==0.5.1
